Vinyl.CoRec

Description

Documentation

data CoRec :: (k -> *) -> [k] -> * where

Generalize algebraic sum types.

Constructors

Col :: t ∈ ts => !(f t) -> CoRec f ts

type OneOf = CoRec Identity

a value of type OneOf [x,y,z] holds a value of one of the types, either x or y or z.

newtype Outputs b a

A function type constructor that takes its arguments in the reverse order.

without type lambdas, we need a type constructor to reorder type parameters.

Constructors

O

Fields

runOutputs :: a -> b

type (<--) = Outputs

(b <-- a) is isomorphic to (a -> b).

type Handlers f b as = Rec (Handler f b) as

a product of handlers.

newtype Handler f b a

a flipped CoKleisli.

Constructors

H

Fields

runHandler :: f a -> b

column :: t ∈ ts => t -> OneOf ts

helper to build a OneOf

outputs2handler :: Outputs b a -> Handler Identity b a

Handler generalizes Outputs.

fromUnitCoRec :: CoRec f `[a]` -> f a

inverse with toUnitCoRec

toUnitCoRec :: f a -> CoRec f `[a]`

inverse with fromUnitCoRec

handle :: Rec (Handler f b) as -> CoRec f as -> b

you consume a coproduct with a product of consumers i.e. you must handle every case.

generalizes match.

unitHandlers :: Rec (Handler f (f a)) `[a]`

match :: Rec (Outputs b) as -> OneOf as -> b

perform "pattern matching" upon a CoRec.

a generalized, uncurried, newtyped version of:

either :: (a1 -> b) -> (a2 -> b) -> Either a b

e.g.

>>> :set +m
>>> import Data.Vinyl

truthy is like a final method (i.e. different definitions can be dispatched on a fixed set of different types):

>>> :{
let truthy :: Rec (Outputs Bool) [Int, String, Bool]
    truthy =  (O$ \i -> i==0)
           :& (O$ \s -> s=="")
           :& (O$ \b -> b)
           :& RNil
:}

>>> match truthy (column True       :: OneOf [Int,String,Bool])
True
>>> match truthy (column "nonempty" :: OneOf [Int,String,Bool])
True
>>> match truthy (column (1::Int)   :: OneOf [Int,String,Bool])
True

the integer literal needs a type signature (numeric literals are overloaded) because Haskell can't prove that (Num t, t ∈ [Int,String,Bool]). but it can prove (Int ∈ [Int,String,Bool]). with ExtendedDefaultRules, the type variables can be defaulted. the same problem (and solutions) exists for string literals with OverloadedStrings, and for list literals with OverloadedLists.

type inference does work when the column is immediately consumed@:

>>> match truthy (column False)
False
>>> match truthy (column "")
False
>>> match truthy (column (0::Int))
False